requirements.txt
discord.py>=2.3.0
google-generativeai>=0.8.0